+++ package.json
@@ -11,12 +11,14 @@
   },
   "dependencies": {
     "@faker-js/faker": "^8.4.1",
+    "@hello-pangea/dnd": "^16.6.0",
     "@tanstack/react-query": "^5.36.0",
     "@tanstack/react-query-devtools": "^5.36.0",
     "axios": "^1.6.8",
     "date-fns": "^3.6.0",
     "eslint-plugin-import": "^2.29.1",
     "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.1",
     "react-calendar": "^5.0.0",
     "react-datepicker": "^6.9.0",
     "react-dom": "^18.2.0",
@@ -26,6 +28,7 @@
   },
   "devDependencies": {
     "@types/react": "^18.2.66",
+    "@types/react-beautiful-dnd": "^13.1.8",
     "@types/react-datepicker": "^6.2.0",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
