+++ src/components/modal/modal-header.tsx
@@ -11,9 +11,9 @@ const ModalHeader = ({ headerTitle, onClose }: ReportHeaderProps) => {
   return (
     <S.HeaderContainer>
       <S.ModalTitle>{headerTitle}</S.ModalTitle>
-      <S.CloseButton onClick={onClose}>
-        <S.CloseIcon src={xCircle} alt="x" />
-      </S.CloseButton>
+      <button onClick={onClose}>
+        <img src={xCircle} alt="x" />
+      </button>
     </S.HeaderContainer>
   );
 };
@@ -32,8 +32,6 @@ const S = {
     line-height: 36px;
     font-weight: 900;
   `,
-  CloseButton: styled.button``,
-  CloseIcon: styled.img``,
 };
 
 export default ModalHeader;
