+++ src/pages/group-home/components/report-modal/report-modal.tsx
@@ -54,7 +54,6 @@ const S = {
       'member chat'
       'time chat';
     max-width: 962px;
-    width: 100%;
     gap: 20px;
     margin: 10px;
   `,
